fix: serialize BreakdownItem value/weight as float not Decimal
Pydantic serializes Decimal as strings which Recharts cannot use for
calculating pie slice proportions, causing blank donut charts.
BreakdownItem uses float since these are display-only chart values.

+++ backend/app/schemas/portfolio.py
@@ -35,8 +35,8 @@ class PortfolioValuePoint(BaseModel):
 
 class BreakdownItem(BaseModel):
     name: str
-    value: Decimal
-    weight: Decimal
+    value: float
+    weight: float
     holdings_count: int
